Allow redirect url to be specified
Cleans up authentication to coalesce into single redirect url. This is
stored in the oauth state which is returned by the provider. The reason
we do this is so the provider can just allow /callback/ and we can
customize what redirects are allowed directly from the API.

This also allows for easier local testing by having localhost:port/auth/
be a valid redirect.